fix tether

+++ Splatoon/ObjectInfo.cs
@@ -16,6 +16,6 @@ namespace Splatoon
         internal bool IsChar = false;
         internal float HitboxRadius = 0;
         internal float Distance = 0;
-        internal long Life = 0;
+        internal float Life = 0;
     }
 }
